feat: admin menu for mobile
